+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Issue# 174: Add robot configuration in apps
+
 ## [0.5.1] 2023-01-20
 
 ### Added